refactoring + add sending info about client's server change
